% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinesamples.R
\name{combine.all.sets}
\alias{combine.all.sets}
\title{Combine the data from all sets of fcs files into combined files}
\usage{
combine.all.sets(inpath = getwd(), outpath = getwd(), filenames = NULL)
}
\arguments{
\item{inpath}{The folder/directory where the fcs files are located}

\item{outpath}{The folder/directory where the combined fcs file
will be written to}

\item{filenames}{A data.frame which maps filenames to
basenames. This data.frame needs two columns, one called
`filename`, and another `basename`, where each row contains
the name of the file in `filename`, and the name of the set it
belongs to in `basename`; all the files with the same
`basename` will be combined into one file `basename.fcs`.}
}
\value{
This function does not return anything. Instead, it
    creates a set of fcs files combining the events from several
    files with the same basename. Existing files with that name
    will be overwritten without warning.
}
\description{
Combine the data from all sets of fcs files that come from
repeated runs of samples from the same patient and under the same
conditions, and write a set of combined files to disk. Neither the
flowSet, nor the combined flowFrame for any of the sets are kept
in memory after writing the combined flowFrame to disk, though all
the data for a single set needs to be kept in memory to combine
it.
}
\details{
The most simple usage would be simply as combine.all.sets(), to
automatically combine files in the current working directory. This
is equivalent to combine.all.sets(filenames=get.filenames()). To
do the same in another folder, one can do
e.g. combine.all.sets(path="../fcsFiles/"), which is equivalent
again to combine.all.sets(path="../fcsFiles/",
filenames=get.filenames(path="../fcsFiles/")).

To be able to know which files belong to each other, this function
either calls `get.filenames`, or (if \code{!is.null(filenames)})
relies on a data.frame which maps filenames to basenames. This
data.frame needs two columns, one called `filename`, and another
`basename`, where each row contains the name of the file in
`filename`, and the name of the set it belongs to in `basename`;
all the files with the same `basename` will be combined into one
file `basename.fcs`. For an example, see the output of
`get.filenames`, when each file in a directory ends with something
like ` 1_C04.fcs`, that is a space, a number, an underscore, the
place on a plate and ".fcs" to indicate the n-th (here 1st) sample
in a set. See the help for `get.filenames` for more details.
}
