% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinesamples.R
\name{combine.set}
\alias{combine.set}
\title{Combine the data from a set of fcs files into one combined file}
\usage{
combine.set(filenames, basename, inpath = getwd(), outpath = getwd())
}
\arguments{
\item{filenames}{A character vector containing filenames for each
of the fcs files to combine. This is the order in which the
data will be written to the combined fcs file.}

\item{basename}{The filename of the new fcs file, without ".fcs"}

\item{inpath}{The folder/directory where the fcs files are located}

\item{outpath}{The folder/directory where the combined fcs file
will be written to}
}
\value{
This function just returns `basename`, it does *not*
    return any flowFrame or flowSet.
}
\description{
Combine the data from a set of fcs files that come from repeated
runs of samples from the same patient and under the same
conditions, and write a combined file to disk. Neither the
flowSet, nor the combined flowFrame are kept in memory after
writing the combined flowFrame to disk, though all the data needs
to be kept in memory to combine it.
}
