\name{fcscat-package}
\alias{fcscat-package}
\alias{fcscat}
\docType{package}
\title{\packageTitle{fcscat}}
\description{\packageDescription{fcscat}}
\section{Package Content}{\packageIndices{fcscat}}
\author{\packageAuthor{fcscat}}
\section{Maintainer}{\packageMaintainer{fcscat}}
\keyword{package}
